Refactore stories structure. Align Grid with the layout structure. New doc for storybook
