+++ packages/style/index.scss
@@ -1,3 +1,2 @@
-@import "./cards/grid_card";
-@import "./cards/grid_card";
+@import "./view/cards/grid";
 @import "./layout/grid";
