+++ packages/style/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@ferlab/style",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "description": "Core components for scientific research data portals",
   "publishConfig": {
     "access": "public"
